builder: allow multiple target directories for the same release
kernel_dirs has so far been treated as a dictionary with the release
as the key, and the target directory as the value.
However, for ubuntu we have the need to build the exact same release
onto multiple target directories (i.e. package versions).
Rework the dictionary as a list of tuples.