fix eBPF build for Fedora 35
Re-add clang and llvm to the builder container for Fedora 35,
which were initially removed because at the time they would
generate code that would not pass the verifier test upon loading.

The version now present in the image does not seem to suffer from
this problem anymore, so re-add them.

This reverts commit 66724e2f0aa069f799f9165cae8d4b9e8bbfa9c8.