optimize download
In order to handle incomplete downloads, the current implementation
will always try to append any remaining bytes to an already existing files.
But this means issuing at least one HTTP request per file, which might
slow down the process dramatically.
Instead, use temporary .part files and rename to the their filename
once complete -- this way we can assume that a file with the right name
is always complete and we can skip the download altogether.