draft: handle new driver's parameter naming
the new driver with semantic versioning does not currently
compile because cmake parameter names have been renamed.
So in order to have something working with the existing naming
scheme, use the new parameter names with the old semantics.
Which I have to admin, feels like defeating the whole purpose
of the renaming. But at least we'd get probes for the next release.