Dedicated AmazonLinux2 builder (#51)
* Dedicated AmazonLinux2 builder

AmazonLinux2 ships with gcc 7.3, which makes us choose
the builder image based on Fedora 27. That comes with clang 5,
which does build the eBPF probe successully but the probe
fails verification, e.g. with kernel 4.14.275-207.503.amzn2.

Create a dedicated AmazonLinux2 builder, with clang 7 (as it's
more battle tested than clang 11 which comes in the default
clang package on AmazonLinux2).

* Add a comment about using clang-7 over clang-11 in amzn2 builder

Co-authored-by: Gerlando Falauto <gerlando.falauto@sysdig.com>