parallel builds: mount driver code readonly and run cmake for ebpf (#52)
this will be the preliminary step to parallel builds:

====
- part 1: mount code readonly
====
- part 2: fix handling of bpf probes when source is mounted readonly

the previous implementation, with the driver folder mounted in r/w,
was working through some side effects of building the kmod:
it would create files (e.g. driver_config.h) which would be
available to the subsequent invocation for eBPF.
Now that such folders are no longer shared through the host filesystem,
they would get the artifacts from whatever build was done on the host
prior to introducing the parallel builders.
The end result was that we would get probes built with the wrong
version information.

So fix this by having an explicit run of cmake on it for the
eBPF case, too.
Also, it would be a good idea to run the builder off a clean
git checkout for the driver.