experimental: add qemu-user-static to centos-gcc9.2
When running CentOS 7 as the host OS, it looks like we can't
just register qemu from within a container and be done with it.
It looks as if with that OS/kernel, binfmt tries to resolve your
qemu binary as you try to run your executable, from within its
context, so that you need to have qemu in your running container.
So add it.