allow for driver sources coming from agent-kmodule.tgz
The current procedure is based on the assumption of
having a full checkout of agent-libs.
However, the source code (for both kernel module and eBPF
probe) is also available as the draios-agent-kmodule.tar.gz
file where all parameters are already populated.
Enable that scenario while also retaining compatibility with
the existing one.
Notice the unavoidable tweaks since there is no longer
a driver/ directory available.