[SMAGENT-5267] Fix prebuilt builders for air-gapped environments
When introducing support for cross-architectural builds,
we had to add a suffix specifying the architecture
(e.g. ubuntu-gcc11.2-bpf-amd64) in order to tell the
multiple builders apart.

The step that prebuilds all the builders to be imported
into the air-gapped host (-P) is currently implemented
within the bash entrypoint and does not add that suffix,
thereby invalidating the whole procedure.

So defer the actual prebuilding to a newly-introduced
'prebuild' command of the python code, so that the whole
naming logic is encapsulated within the same code for
both producer (prebuild) and consumer (actual building).