ubuntu: add a jammy builder with gcc 11.4
gcc-11 11.4 is the required compiler for 6.2.0-1009
(backported to jammy), but we currently only have
builders marked for gcc 11.2 and 11.3 (see #102).

As it was previously noted, this results in the builder
for gcc12.2 being picked as the next best choice,
which fails because there is no gcc-11 binary.

So along the line of what was already done, add a
dedicated builder for gcc11.4.