feat: introduce driver/kernel yaml blacklist
introduce an external blacklist, i.e. a way to specify a set of
kernel releases that should be skipped for a given set of
driver versions.
The set of releases is specified through a regex and a
conditional expression on the exploded values, expressed
as a Jinja2 template.
The most notable use case is an older driver version which
is incompatible with newer kernels due to compiler issues.