[SMAGENT-6644] Handle zstd compression
Starting from Ubuntu 23.10, the default compression
method for kernel modules is zstd.
When running the agent-kmodule container on such
host distros, one of two things might happen:

1) If zstd is not installed in the container,
compression will fail:

/usr/sbin/dkms: line 1127: zstd: command not found

This will result in the untouched .ko file
being present under /var/lib/dkms.
So we will find the original .ko file and
the probe loader will be able to install it by
manually picking the path under /var/lib/dkms.

2) If zstd is indeed installed, dkms will destroy
the original .ko file so we won't be able to install
it.

Add the necessary plumbing to get the script to
work in the second case.