fix(rhel): honor kernel filtering for local distros
Today if we pass a kernel filter (-f/--kernel-filter)
e.g. 5.14.0-427.18.1.el9_4 to a local distro crawler
like CustomCentos, it is completely ignored so all
kernels are built instead.

This change will make the filter start being honored
(treating it as a prefix so that 5.14.0 can also be used).

Also, make the script return an error if the list of
matched kernels is empty.