fix(ubuntu): add builder with gcc13.3
Recent ubuntu noble kernels (namely: 6.8.0-50-generic)
are being compiled with gcc13.3, which is only available
through noble-proposed.
So add the appropriate builder for it.

While at it, fix the gcc13.2 builder (currently based
on mantic, already out of date) by making it a clone
of 13.3, but with the gcc version pinned to 13.2 instead.