[SMAGENT-9129] Implement kernel cache
Implement a rudimentary kernel cache that would easily reduce crawling
efforts. Logic is as follows.

A crawl is bound to a specific query, that is: kernel type (aka distro),
architecture, distro filter, kernel filter.

Whenever an actual crawl is successful, save it into the cache
file together with its timestamp.

Whenever a crawl is requested, look up the cache to see if a resultset
is already present with a timestamp no older than 20 hours -- just to
make sure we perform a crawl at least once a day. If one is available,
use it -- otherwise, just proceed as usual with a regular crawl and
cache the resultset to be reused at the next iteration.